repository e service reativo

+++ src/main/java/domain/Agencia.java
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 public class Agencia {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String nome;
 
@@ -21,7 +21,7 @@ public class Agencia {
 
     public Agencia() { }
 
-    public Agencia(Integer id, String nome, String razaoSocial, String cnpj, Endereco endereco) {
+    public Agencia(Long id, String nome, String razaoSocial, String cnpj, Endereco endereco) {
         this.id = id;
         this.nome = nome;
         this.razaoSocial = razaoSocial;
@@ -29,7 +29,7 @@ public class Agencia {
         this.endereco = endereco;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
